Move graphql operations to separate folder in frontend
